requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
gunicorn==19.7.1
markdown2==2.3.4
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
